Update README.rst
Update readme description about the results of the roundtrip event. The notion of inverse functions captures the dynamics happening here. In general for functions f: X -> Y and g: Y -> X, we have that f∘g: Y -> Y and g∘f: X -> X. So commutativity would only make sense if the sets X and Y were the same. In this case we have encodebytes: bytes -> strings and decodebytes: strings -> bytes. Formally, these can't commute as the sets are different but each is indeed an inverse to the other.